+++ src/main/java/com/example/demo/model/entity/demo/ProductTest.java
@@ -3,6 +3,8 @@ package com.example.demo.model.entity.demo;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
+
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.io.Serializable;
 
@@ -29,7 +31,7 @@ public class ProductTest implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private BigInteger id;
 
     private String guid;
